Input final re-factor
