+++ src/components/Atoms/Controls/Input/index.tsx
@@ -1,16 +1,15 @@
 import * as React from "react";
 import { useFormContext, RegisterOptions, useWatch } from "react-hook-form";
-import { tv } from "tailwind-variants";
 import Switch from "@/components/Atoms/Controls/Switch";
 import DropdownMenu from "@/components/Molecules/Dropdowns";
 import Radio from "../RadioButton";
 import { MenuItem, RadioOptions } from "@/types";
 import { IoIosEyeOff } from "react-icons/io";
 import { IoEye } from "react-icons/io5";
-import { Tooltip } from "../../Misc/Tooltip";
-import { AiOutlineInfoCircle } from "react-icons/ai";
-import Badge from "../../Misc/Badge";
-interface CustomChangeEvent {
+import { InputLabel } from "./components/InputLabel";
+import { InputError } from "./components/InputError";
+import DefaultInput from "./components/DefaultInput";
+export interface CustomChangeEvent {
   target: {
     name: string;
     value: any;
@@ -18,7 +17,7 @@ interface CustomChangeEvent {
   };
 }
 
-type InputType =
+export type InputType =
   | "text"
   | "email"
   | "password"
@@ -47,27 +46,6 @@ interface CustomInputProps
     event: React.ChangeEvent<HTMLInputElement> | CustomChangeEvent,
   ) => void;
 }
-
-const inputClass = tv({
-  base: "body-3 w-full rounded-xl bg-white bg-opacity-10 ring-0 outline-0 focus:ring-1 active:ring-1 focus:ring-white active:ring-white transition-all ease-in-out duration-300 text-white",
-  variants: {
-    error: {
-      true: "border border-error-600",
-      false: "",
-    },
-    size: {
-      sm: "body-5 font-medium rounded-md p-2",
-      md: "body-3 rounded-xl p-3",
-      lg: "body-3 rounded-xl p-4",
-    },
-    withIcon: {
-      true: "pr-12", // Add padding to the right to accommodate the icon
-    },
-    disableNumberInputDefaults: {
-      true: "[&::-webkit-outer-spin-button]:appearance-none [&::-webkit-inner-spin-button]:appearance-none [&::-webkit-inner-spin-button]:m-0 [&::-webkit-outer-spin-button]:m-0 [appearance:textfield]",
-    },
-  },
-});
 
 const Input = React.forwardRef<HTMLInputElement, CustomInputProps>(
   (
@@ -286,123 +264,27 @@ const Input = React.forwardRef<HTMLInputElement, CustomInputProps>(
       }
     };
 
-    const InputLabel = () => {
-      if (!label && !badge && !tooltip) return null;
-      return (
-        <div className="flex flex-row items-center space-x-3">
-          {label && type !== "switch" && type !== "radio" && (
-            <label htmlFor={name} className="body-2 text-white">
-              {label}
-            </label>
-          )}
-          {badge && (
-            <Badge
-              variant="neutral"
-              rounded={true}
-              label={badge}
-              caption={true}
-            />
-          )}
-          {tooltip && (
-            <Tooltip content={tooltip} placement={tooltipPlacement}>
-              <div>
-                <AiOutlineInfoCircle
-                  className="text-white flex-shrink-0"
-                  size={16}
-                />
-              </div>
-            </Tooltip>
-          )}
-        </div>
-      );
-    };
-
-    const InputError = () => {
-      if (!disabled && !errors[name]) return null;
-      return (
-        <span className="text-error-600 text-xs">
-          {errors[name]?.message as string}
-        </span>
-      );
-    };
-
     React.useEffect(() => {
       if (!disabled) resetField(name, { defaultValue: props.defaultValue });
     }, [disabled]);
 
     return (
       <div
-        className={`flex flex-col gap-2 w-full py-1 h-fit ${disabled ? "opacity-40" : "opacity-100"}`}
+        className={`flex flex-col gap-2 w-full  py-1 h-full ${disabled ? "opacity-40" : "opacity-100"}`}
       >
-        <InputLabel />
+        <InputLabel
+          name={name}
+          label={label}
+          tooltip={tooltip}
+          tooltipPlacement={tooltipPlacement}
+          type={type}
+        />
         {renderInput()}
-        <InputError />
+        <InputError error={errors[name]?.message as string} />
       </div>
     );
   },
 );
-
-const DefaultInput = ({
-  defaultValue,
-  disabled,
-  isError,
-  type,
-  withIcon,
-  handleChange,
-  inputRef,
-  className,
-  parentRef,
-  children,
-  disableNumberInputDefaults,
-  size,
-  ...restProps
-}: {
-  defaultValue?: string | number | readonly string[] | undefined;
-  disabled?: boolean;
-  isError?: boolean;
-  type: InputType;
-  handleChange?: (
-    event: React.ChangeEvent<HTMLInputElement> | CustomChangeEvent,
-  ) => void;
-  inputRef: (e: HTMLInputElement | null) => void;
-  className?: string;
-  withIcon?: boolean;
-  parentRef: React.ForwardedRef<HTMLInputElement>;
-  children?: React.ReactNode;
-  restProps?: any;
-  disableNumberInputDefaults?: boolean;
-  size?: "sm" | "md" | "lg";
-}) => {
-  return (
-    <div className="relative w-full">
-      <input
-        {...restProps}
-        defaultValue={defaultValue}
-        className={inputClass({
-          error: isError,
-          withIcon: withIcon,
-          disableNumberInputDefaults: disableNumberInputDefaults,
-          className,
-          size,
-        })}
-        disabled={disabled}
-        id={randomIdGenerator()}
-        type={type}
-        onChange={handleChange}
-        ref={(e) => {
-          if (inputRef) inputRef(e);
-          if (typeof parentRef === "function") parentRef(e);
-          else if (parentRef) parentRef.current = e;
-        }}
-      />
-      {children}
-    </div>
-  );
-};
-
-const randomIdGenerator = () => {
-  return Math.random().toString(36).substring(2, 15);
-};
 
 Input.displayName = "Input";
